feat: 세팅 페이지에서 color selector 분리
