+++ ReactNativeStudy/Zero/Week1/Todo/src/pages/Setting.tsx
@@ -1,15 +1,13 @@
 import { StyleSheet, Text, View } from 'react-native';
 import { theme } from '../constants';
-import Icon from '../components/Icon';
 import React from 'react';
-import { useColor } from '../store/colorState';
 import PageTitle from '../components/PageTitle';
+import ColorSelector from '../components/ColorSelector';
 
 const colors: (keyof typeof theme.color)[] = ['Orange', 'Green', 'Blue', 'Purple', 'Pink'];
 
 function Setting() {
-    const selectedColor = useColor().color;
-    const handleChangeSelectedColor = useColor().changeColor;
+
 
     return (
         <View style={style.container}>
@@ -17,15 +15,7 @@ function Setting() {
             <Text style={style.text}>색상을 선택해주세요.</Text>
             <View style={style.colorWrapper}>
                 {colors.map((color) => (
-                    color === selectedColor ?
-                        <Icon width="30" height="30" type="check" fill={theme.color[color]} key={color} />
-                        : <View style={{
-                            ...style.selectedColor,
-                            backgroundColor: theme.color[color],
-                        }}
-                            key={color}
-                            onTouchEnd={() => handleChangeSelectedColor(color)}
-                        />
+                    <ColorSelector color={color} key={color} />
                 ))}
             </View>
         </View>
@@ -50,11 +40,6 @@ const style = StyleSheet.create({
         alignItems: 'center',
         paddingHorizontal: 66,
         width: '100%',
-    },
-    selectedColor: {
-        width: 24,
-        height: 24,
-        borderRadius: 100,
     },
 });
